Added Initial File
New File for the network monitor.

Features used from original example file:
Display Hostname
Display IPaddr

Additional Features:
Instead of displaying wifi, this monitor pings google to check that my DNS server (in this case another raspberry pi running pihole) is up.

If the device is up it will display DNS is UP and turn the LED green.

If not, it will display DNS is DOWN and turn the LED Red.